requirements.txt
rasa
rasa-sdk
tensorflow
sqlalchemy<2.0
psycopg2-binary